xmlui: Enable opt-in popup for GDPR
If a user agrees to the privacy policy we will enable Google Analytics,
otherwise we will set a property to disable Google Analytics. Currently
it seems there is a bug with something in the DSpace theme.js bundle or
something that causes the Agree / Disagree buttons to not dismiss after
the user clicks them. This is annoying, but the cookies are indeed set
so I don't care for now.

Also, even if a user agrees to the policy Google Analytics still won't
actually run until the next page load. I didn't have time to figure out
how to queue that up.

Also also, there has to be a better way of doing this that doesn't need
me to copy it to each theme, but I don't have time. Ideally it would be
an XSLT template.

See: https://developers.google.com/analytics/devguides/collection/analyticsjs/user-opt-out