Re-order Navigation in right sidebar
We override dspace-xmlui's Navigation.java so that we can place the
"My Account" and Discovery links higher than Browse, etc. It seems
there is no other way to do this other than simply overriding it in
the code!

I need to remember that I did this so that I check for changes to
DSpace's own Navigation.java so I can port them to our copy.