dspace/modules/xmlui-mirage2/pom.xml: Don't copy node_modules
This copies the node_modules directory for *every* theme into the
installation directory. For twenty-six themes this means an extra
200,000 files or so, which increases the time to install, and is
much heavier on the disk.

I modified the grunt tasks to bundle some of these into theme.js
and copy others directly to the scripts directory so we don't need
node_modules anymore.