Re-work Mirage 2 build to use Node.js libraries
This completely removes the Mirage 2 reliance on Ruby sass and the
horribly old compass. Firstly this allows us to eliminate the need
for copying a node_modules directory to each of our 25+ themes, a
savings of around 250,000 files that no longer need to be copied to
the installation directory.

In order to use a more modern SASS processor we need to import the
old compass mixins from compass-sass-mixins on npm and adjust some
of our .scss files for new semantics:

  - styles/_style.scss: explicitly import main.scss to avoid ambig-
    uity over main.scss vs main.css when importing from 0_CGIAR
  - styles/_style.scss: add override for breadcrumb Glyphicon that
    would otherwise be rendered as "/Â" due to the SASS processor
    converting the ASCII escape code to an invalid Unicode literal
  - styles/classic_mirage_color_scheme/_main.scss: import jquery-ui
    without its .css extension so that it is included in its entir-
    ety instead of as an "@import url()" statement, which inlines
    one more CSS file into our main.css and causes one less network
    request in XMLUI themes

Furthermore, we modify the font paths for Bootstrap Glyphicon and
Font Awesome to point to a new top-level "fonts" directory that is
populated during the grunt build process (so we don't need to refe-
rence them in node_modules at run time).

Lastly, we apply a build-time hack to replace a hard-code reference
to compass mixins in Atmire's styles/_statlets.scss.

Eventually it would be nice to move to dart-sass, which is the new
official SASS pre-processor implementation, but it currently only
outputs UTF-8 and this garbles our Glyphicons because I guess some
of our CSS/SCSS/HTML is not in UTF-8...?