xmlui: Use Font Awesome SVG via JS instead of CSS
By using Font Awesome 5 SVGs via JavaScript we reduce the amount of
data the client downloads when viewing the site by about 105K. One
additional benefit is that we make two less network requests due to
not needing the web fonts themselves.

To make this work we need to modify the Grunt tasks to use webpack
to package the Font Awesome JavaScript from ES6 to UMD because the
UglifyJS environment only supports ES5 (ie, no modules or require).

There are minor tweaks to the stylesheets needed for the SVG icons.

See: https://fontawesome.com/how-to-use/on-the-web/advanced/svg-javascript-core