xmlui: Add missing Font Awesome icons used by Atmire
Atmire's CUA is using Font Awesome 4 and some of the icon names are
the same, though we need to import the v4 shims for two. There are
unfortunately two others that are not used as spans, but instead as
Unicode characters, that we can't replace this way:

  - f06e
  - f019

These are the "eye" and "download" icons used in statlets.