input-forms.xml: Use title case for countries
We decided that this makes it easier to compare country data across
repositories in tools like the AReS Explorer without having to make
a bunch of mappings. Also, this is actually how they appear in the
ISO 3166-1 controlled vocabulary anyways.

I used a kinda crazy vim regex with a "negative lookahead" to edit
the values directly in the file:

    :'<,'>s/\<\(pair\|displayed\|stored\|value\|AND\)\@!\(\w\)\(\w*\|\)\>/\u\2\L\3/g

It searches for words that are not part of the XML tags themselves
like "pair" and "displayed-value" and upper cases the first letter
and lowercases the rest using groups and backreferences.