pom.xml: Version bump for Atmire CUA
This fixes some issues with the Atmire Content and Usage Analysis
module. Most importantly it fixes the runaway database locking that
was caused by CUA incorrectly probing the Solr cores and timing out.
Secondly, it updates the duplicates processor that Atmire wrote to
remove duplicate documents from the Solr statistics.